requirements.txt
Django
django-debug-toolbar
django-filter
djangorestframework
djangorestframework-simplejwt
djoser
drf-spectacular
python-dotenv
psycopg[binary]




